Add support for in-kernel headers
This will allow building of kernel modules and running of BCC programs.
The headers are provided in compressed tarball in /proc/kheaders.tbz.

A "kheaders" boot parameter needs to be passed to the boot arguments
otherwise the memory containing the compressed archive will be freed.
There is an increase in kernel binary size from 9MB to 16MB. We can
reduce that further.

This is an RFC patch. Future enhancements will involve making it a
CONFIG option. The RFC just makes it built always.

Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>