Add a test module to test ikheaders feature
This test module can be used to test the ikheaders feature.

Note that this test module is just a part of this RFC for testing
purposes. It should ideally be a kselftest.

Also note that I added some scripts like 'hplist-build.sh' which
actually manually builds a set of minimal headers into a temp directory
and builds the module using them. This script is useful to know what is
really needed.

But if you just try to build the module directly, using make, then you
need to boot a kernel with ikheaders feature. The Makefile will take of
the rest.
cd testmod
make

Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>