link-vmlinux: Make updated compile.h be included in kheaders
compile.h is updated after .version is incremented. However the older
compile.h with older version is stored in the kheaders.tar.xz archive.
This causes an old UTS_VERSION to be stored in the compile.h in the
archive. This patch regenerates the archive if .version is updated thus
causing the compile.h containing the new UTS_VERSION to be stored in the
archive.

Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>