rcu: Report quiescent state in kernel mode for !PREEMPT_RCU systems
For !PREEMPT_RCU configuration, I notice that some times the
force-quiescent state loop (rcu_gp_fqs_loop()) in the GP thread can take
a long time to execute.  It does several loops before finally the
dynticks-idle counters of one of thes last CPUs holding out are updated.

One other opportunity for the loop to terminate is, if the QSes are
reported from the culprit CPUs. This termination happens even if the
dyntick-idle counters of the said CPUs are not updated. Say, if the CPU
in concern was non-RCU-idle and in kernel mode - then its dynticks
counters would remain the same for a while (to represent that the CPU is
not RCU-idle).

Digging further I notice that the CPU holding out are actually in
kernel-mode and are receiving the scheduler-tick but are not reporting
the quiescent state. If they do, then there is a chance for the loop to
terminate early. So, in this patch we make it a little better here by
reporting a QS for !PREEMPT_RCU if we received the tick while the kernel
was not in any preempt-disabled or softirq-disabled section. We are
already doing the same for PREEMPT_RCU anyway.

Doing some testing I noticed that the amount of looping of the
rcu_gp_fqs_loop is reduced with this. Some times the looping is reduced
quite a bit and just 1 or 2 loops are required to accomplish the task.
This likely also makes the GP complete faster and reduces CPU
consumption.

Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>
diff-note-start
Note that I am testing this patch further, but I wanted to post it as an
RFC in advance for discussion/review. I did some basic testing and
everything seems fine so far. Thanks.