Add polling support to pidfd
pidfd are /proc/pid directory referring to a task-group. When the parent
of the task group is to be notified, the tasks waiting on a poll are
awakened.

Android low memory killer needs this to replace code that currently
checks for existence of /proc/pid which is both racy and slow. The pidfd
poll approach is race free, and also allows the low memory killer to do
other things (such as by polling on other fds).

Co-developed-by: Daniel Colascione <dancol@google.com>
Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>
option-start
option-prefix PATCH v
option-skip-checkpatch

diff-note-start
TODO: Address fixups from reviewers for the kselftests.