RFC/INTERNAL-REVIEW: Add LSM hooks for perf_event_open
Yabin, Jeff, could you please take a look at this patch and see if it
would work for SimplePerf?

The idea is to lower the sysctl_perf_event_paranoid to 1 and then use
sepolicy to give access. This helps us give access to perfetto to use
perf_event_open.
----
Actually commit message:

Currently the sysctl_perf_event_paranoid controls accesses to
perf_event. In Android, we have a patch that sets this to 3 which was
attempted to be upstreamed [1] however was rejected and a suggestion was
made instead to use LSM hooks instead of killing world access to perf.

This patch attempts to use LSM hooks in conjunction with the sysctl.

[1] https://patchwork.kernel.org/patch/9249919/
Cc: yabinc@google.com
Cc: jeffv@google.com
Cc: primiano@google.com
Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>
option-private-rfc
option-start
option-end